Add: tus-endpoint now only needs upload-token and moved to Metadata
When using web-frontends around this API, two problems had to be
solved when uploading to tus:

1) tus needed the api-token, but the web-frontend normally doesn't
show this (as it is between the web-frontend server and the API;
the client has nothing to do with this). So switching to upload-token
only solves this problem. This is also security-wise safe, as the
upload-token is a secret generated on-demand. And it only lives for
roughly 15 minutes.

2) CORS is implemented, correctly, rather strict in tusd. This means
that we cannot use additional headers outside the ones in the CORS
header. This means that the upload-token has to move from headers
to Metadata. This also means that we have to trust the information
on disk about the upload. This too is security-wise safe, as this
file-uuid is a secret generated on-demand.

This commit combines these two findings and solves them both.