Fix: change the login flow to an OAuth2 PKCE flow (#15)
This avoids other websites being able to steal API tokens and
modify any content on the API, without the user knowing.

If people want to register their application, they can make a
pull-request. This follows OAuth2 design, although we don't have
a dynamic portal to register your app, but it needs to be done
via a pull-request.