Add #22: allow uploading .zip / .tar / .tar.gz additional to individual files
These archives are extracted and the individual files are added
to the list. Folder-structure is maintained.

The most common way to make tar-balls is to have a single folder
as root-folder. Therefor, this folder is skipped (but only if
there is exactly one folder and no files in the root).